[Issue-3556] Update UI
