+++ packages/extension-base/src/utils/account/transform.ts
@@ -14,8 +14,6 @@ import { SingleAddress, SubjectInfo } from '@subwallet/ui-keyring/observable/typ
 import { hexStripPrefix, u8aToHex } from '@polkadot/util';
 import { blake2AsHex, mnemonicToEntropy, mnemonicValidate } from '@polkadot/util-crypto';
 
-import { getDerivationInfo } from './derive';
-
 export const createAccountProxyId = (_suri: string, derivationPath?: string) => {
   let data: string = _suri;
 
@@ -88,7 +86,7 @@ export const getAccountSignMode = (address: string, _meta?: KeyringPair$Meta): A
   }
 };
 
-export const getAccountActions = (signMode: AccountSignMode, networkType: AccountChainType, type: KeypairType, _meta?: KeyringPair$Meta, parentAccount?: AccountJson): AccountActions[] => {
+export const getAccountActions = (signMode: AccountSignMode, networkType: AccountChainType, type: KeypairType, _meta?: KeyringPair$Meta): AccountActions[] => {
   const result: AccountActions[] = [];
   const meta = _meta as AccountMetadataData;
 
@@ -117,22 +115,12 @@ export const getAccountActions = (signMode: AccountSignMode, networkType: Accoun
   }
 
   // Derive
-  if (signMode === AccountSignMode.PASSWORD) {
-    const deriveInfo = getDerivationInfo(type, meta as KeyringPair$Meta);
-
+  if (signMode === AccountSignMode.PASSWORD && !meta.parentAddress) {
     if (networkType === AccountChainType.SUBSTRATE) {
-      if (deriveInfo.depth < 2) {
-        result.push(AccountActions.DERIVE);
-      }
+      result.push(AccountActions.DERIVE);
     } else if (type !== 'ton-native') {
-      if (deriveInfo.depth === 0) {
-        if (meta && meta.isMasterAccount) {
-          result.push(AccountActions.DERIVE);
-        }
-      } else if (deriveInfo.depth === 1) {
-        if (parentAccount && parentAccount.accountActions.includes(AccountActions.DERIVE)) {
-          result.push(AccountActions.DERIVE);
-        }
+      if (meta && meta.isMasterAccount) {
+        result.push(AccountActions.DERIVE);
       }
     }
   }
@@ -350,7 +338,7 @@ export const getAccountTokenTypes = (type: KeypairType): _AssetType[] => {
  * 3. If chain information is provided, add full data (accountActions, transactionActions) to account.
  * 4. Returns an `AccountJson` object containing the transformed account data.
  */
-export const transformAccount = (address: string, _type?: KeypairType, meta?: KeyringPair$Meta, chainInfoMap?: Record<string, _ChainInfo>, parentAccount?: AccountJson): AccountJson => {
+export const transformAccount = (address: string, _type?: KeypairType, meta?: KeyringPair$Meta, chainInfoMap?: Record<string, _ChainInfo>): AccountJson => {
   const signMode = getAccountSignMode(address, meta);
   const type = _type || getKeypairTypeByAddress(address);
   const chainType: AccountChainType = getAccountChainType(type);
@@ -379,7 +367,7 @@ export const transformAccount = (address: string, _type?: KeypairType, meta?: Ke
     }
   }
 
-  const accountActions = getAccountActions(signMode, chainType, type, meta, parentAccount);
+  const accountActions = getAccountActions(signMode, chainType, type, meta);
   const transactionActions = getAccountTransactionActions(signMode, chainType, type, meta, specialChain);
   const tokenTypes = getAccountTokenTypes(type);
 
@@ -400,15 +388,13 @@ export const transformAccount = (address: string, _type?: KeypairType, meta?: Ke
 
 export const singleAddressToAccount = (
   { json: { address, meta }, type }: SingleAddress,
-  chainInfoMap?: Record<string, _ChainInfo>,
-  parentAccount?: AccountJson
-): AccountJson => transformAccount(address, type, meta, chainInfoMap, parentAccount);
+  chainInfoMap?: Record<string, _ChainInfo>
+): AccountJson => transformAccount(address, type, meta, chainInfoMap);
 
 export const pairToAccount = (
   { address, meta, type }: KeyringPair,
-  chainInfoMap?: Record<string, _ChainInfo>,
-  parentAccount?: AccountJson
-): AccountJson => transformAccount(address, type, meta, chainInfoMap, parentAccount);
+  chainInfoMap?: Record<string, _ChainInfo>
+): AccountJson => transformAccount(address, type, meta, chainInfoMap);
 
 export const transformAccounts = (accounts: SubjectInfo): AccountJson[] => Object.values(accounts).map((data) => singleAddressToAccount(data));
 
@@ -581,44 +567,9 @@ export const _combineAccounts = (accounts: AccountJson[], modifyPairs: ModifyPai
  * @summary Use for `AccountContext` subscribe account
  * */
 export const combineAccountsWithSubjectInfo = (pairs: SubjectInfo, modifyPairs: ModifyPairStoreData, accountProxies: AccountProxyStoreData, chainInfoMap?: Record<string, _ChainInfo>) => {
-  const accountMap = {} as Record<string, AccountJson>;
+  const accounts = Object.values(pairs).map((data) => singleAddressToAccount(data, chainInfoMap));
 
-  const convertAccount = (address: string): AccountJson => {
-    const value = pairs[address];
-
-    if (!value) {
-      throw new Error(`Unable to retrieve account '${address}'`);
-    }
-
-    const { json: { meta } } = value;
-
-    if (accountMap[address]) {
-      return accountMap[address];
-    }
-
-    const { parentAddress } = meta as AccountMetadataData;
-    let parentAccount: AccountJson | undefined;
-
-    if (parentAddress) {
-      parentAccount = accountMap[parentAddress];
-
-      if (!parentAccount && pairs[parentAddress]) {
-        parentAccount = convertAccount(parentAddress);
-      }
-    }
-
-    const rs = singleAddressToAccount(value, chainInfoMap, parentAccount);
-
-    accountMap[address] = rs;
-
-    return rs;
-  };
-
-  for (const address of Object.keys(pairs)) {
-    convertAccount(address);
-  }
-
-  return _combineAccounts(Object.values(accountMap), modifyPairs, accountProxies);
+  return _combineAccounts(accounts, modifyPairs, accountProxies);
 };
 
 export const combineAccountsWithKeyPair = (pairs: KeyringPair[], modifyPairs?: ModifyPairStoreData, accountProxies?: AccountProxyStoreData, chainInfoMap?: Record<string, _ChainInfo>) => {
