+++ packages/extension-koni-ui/src/components/Modal/index.tsx
@@ -1,17 +1,19 @@
 // Copyright 2019-2022 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+export { default as AddNetworkWCModal } from './AddNetworkWCModal';
+export { default as DeriveAccountActionModal } from './DeriveAccountActionModal';
 export { default as DisconnectExtensionModal } from './DisconnectExtensionModal';
+export { default as ReceiveModal } from './ReceiveModalNew';
+export { default as RemindBackupSeedPhraseModal } from './RemindBackupSeedPhraseModal';
+export { default as RemindUpgradeUnifiedAccount } from './RemindUpgradeUnifiedAccount';
+export { default as RemindUpgradeVersionModal } from './RemindUpgradeFirefoxVersion';
 export { default as RequestCameraAccessModal } from './RequestCameraAccessModal';
 export { default as RequestCreatePasswordModal } from './RequestCreatePasswordModal';
 export { default as ResetWalletModal } from './ResetWalletModal';
-export { default as RemindBackupSeedPhraseModal } from './RemindBackupSeedPhraseModal';
-export { default as RemindUpgradeVersionModal } from './RemindUpgradeFirefoxVersion';
-export { default as RemindUpgradeUnifiedAccount } from './RemindUpgradeUnifiedAccount';
-export { default as AddNetworkWCModal } from './AddNetworkWCModal';
 export { default as TonWalletContractSelectorModal } from './TonWalletContractSelectorModal';
+
 export { SortingModal } from './SortingModal';
-export { default as ReceiveModal } from './ReceiveModalNew';
 
 export * from './Global';
 export * from './Account';
