+++ packages/extension-koni-ui/src/contexts/WalletModalContextProvider.tsx
@@ -1,13 +1,15 @@
 // Copyright 2019-2022 @polkadot/extension-ui authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { AddressQrModal, AlertModal, AttachAccountModal, ClaimDappStakingRewardsModal, CreateAccountModal, DeriveAccountModal, ImportAccountModal, ImportSeedModal, NewSeedModal, RemindBackupSeedPhraseModal, RequestCameraAccessModal, RequestCreatePasswordModal } from '@subwallet/extension-koni-ui/components';
+import { AddressQrModal, AlertModal, AttachAccountModal, ClaimDappStakingRewardsModal, CreateAccountModal, DeriveAccountActionModal, DeriveAccountListModal, ImportAccountModal, ImportSeedModal, NewSeedModal, RemindBackupSeedPhraseModal, RequestCameraAccessModal, RequestCreatePasswordModal } from '@subwallet/extension-koni-ui/components';
 import { CustomizeModal } from '@subwallet/extension-koni-ui/components/Modal/Customize/CustomizeModal';
-import { ADDRESS_QR_MODAL, EARNING_INSTRUCTION_MODAL, GLOBAL_ALERT_MODAL } from '@subwallet/extension-koni-ui/constants';
+import { AccountDeriveActionProps } from '@subwallet/extension-koni-ui/components/Modal/DeriveAccountActionModal';
+import { ADDRESS_QR_MODAL, DERIVE_ACCOUNT_ACTION_MODAL, EARNING_INSTRUCTION_MODAL, GLOBAL_ALERT_MODAL } from '@subwallet/extension-koni-ui/constants';
 import { useAlert, useGetConfig, useSetSessionLatest } from '@subwallet/extension-koni-ui/hooks';
 import Confirmations from '@subwallet/extension-koni-ui/Popup/Confirmations';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
 import { AlertDialogProps } from '@subwallet/extension-koni-ui/types';
+import { noop } from '@subwallet/extension-koni-ui/utils';
 import { ModalContext, SwModal, useExcludeModal } from '@subwallet/react-ui';
 import CN from 'classnames';
 import React, { useCallback, useContext, useEffect, useMemo, useState } from 'react';
@@ -64,6 +66,9 @@ export interface WalletModalContextType {
   alertModal: {
     open: (props: AlertDialogProps) => void,
     close: VoidFunction
+  },
+  deriveModal: {
+    open: (props: AccountDeriveActionProps) => void
   }
 }
 
@@ -84,6 +89,9 @@ export const WalletModalContext = React.createContext<WalletModalContextType>({
     open: () => {},
     // eslint-disable-next-line @typescript-eslint/no-empty-function
     close: () => {}
+  },
+  deriveModal: {
+    open: noop
   }
 });
 
@@ -111,6 +119,7 @@ export const WalletModalContextProvider = ({ children }: Props) => {
 
   /* Address QR Modal */
   const [addressQrModalProps, setAddressQrModalProps] = useState<AddressQrModalProps | undefined>();
+  const [deriveActionModalProps, setDeriveActionModalProps] = useState<AccountDeriveActionProps | undefined>();
 
   const openAddressQrModal = useCallback((props: AddressQrModalProps) => {
     setAddressQrModalProps(props);
@@ -132,6 +141,13 @@ export const WalletModalContextProvider = ({ children }: Props) => {
 
   /* Address QR Modal */
 
+  /* Derive modal */
+  const openDeriveModal = useCallback((actionProps: AccountDeriveActionProps) => {
+    setDeriveActionModalProps(actionProps);
+    activeModal(DERIVE_ACCOUNT_ACTION_MODAL);
+  }, [activeModal]);
+  /* Derive modal */
+
   const contextValue: WalletModalContextType = useMemo(() => ({
     addressQrModal: {
       open: openAddressQrModal,
@@ -142,8 +158,11 @@ export const WalletModalContextProvider = ({ children }: Props) => {
     alertModal: {
       open: openAlert,
       close: closeAlert
+    },
+    deriveModal: {
+      open: openDeriveModal
     }
-  }), [checkAddressQrModalActive, closeAddressQrModal, closeAlert, openAddressQrModal, openAlert]);
+  }), [checkAddressQrModalActive, closeAddressQrModal, closeAlert, openAddressQrModal, openAlert, openDeriveModal]);
 
   useEffect(() => {
     if (hasMasterPassword && isLocked) {
@@ -195,7 +214,7 @@ export const WalletModalContextProvider = ({ children }: Props) => {
     <AttachAccountModal />
     <NewSeedModal />
     <ImportSeedModal />
-    <DeriveAccountModal />
+    <DeriveAccountListModal />
     <ClaimDappStakingRewardsModal />
     <RequestCreatePasswordModal />
     <RequestCameraAccessModal />
@@ -216,6 +235,14 @@ export const WalletModalContextProvider = ({ children }: Props) => {
         <AlertModal
           modalId={alertModalId}
           {...alertProps}
+        />
+      )
+    }
+
+    {
+      !!deriveActionModalProps && (
+        <DeriveAccountActionModal
+          {...deriveActionModalProps}
         />
       )
     }
